cleanup: move code away from cmd/
